utils/google_drive.py
import os
import json
from datetime import datetime
from google.oauth2 import service_account
from googleapiclient.discovery import build
from googleapiclient.http import MediaFileUpload, MediaIoBaseDownload
import io
import streamlit as st

class GoogleDriveManager:
    def __init__(self):
        self.service = None
        self.root_folder_id = os.getenv('GOOGLE_DRIVE_FOLDER_ID')
        self.initialize_service()
    
    def initialize_service(self):
        """Initialize Google Drive service"""
        try:
            # For now, we'll use a placeholder
            # In production, you'd use service account credentials
            st.info("Google Drive integration requires service account setup")
            self.service = None
        except Exception as e:
            st.error(f"Failed to initialize Google Drive: {e}")
    
    def create_user_folder(self, username):
        """Create a folder structure for a user"""
        if not self.service:
            return self._local_create_folder(username)
        
        # Create main user folder
        user_folder = self._create_folder(username, self.root_folder_id)
        
        # Create subfolders
        subfolders = ['assessments', 'results', 'coaching_sessions', 'uploads']
        for subfolder in subfolders:
            self._create_folder(subfolder, user_folder['id'])
        
        return user_folder['id']
    
    def _create_folder(self, name, parent_id=None):
        """Create a folder in Google Drive"""
        file_metadata = {
            'name': name,
            'mimeType': 'application/vnd.google-apps.folder'
        }
        if parent_id:
            file_metadata['parents'] = [parent_id]
        
        folder = self.service.files().create(
            body=file_metadata,
            fields='id'
        ).execute()
        
        return folder
    
    def save_user_data(self, username, data_type, data):
        """Save user data to Google Drive"""
        if not self.service:
            return self._local_save_data(username, data_type, data)
        
        # Implementation for Google Drive save
        pass
    
    def load_user_data(self, username, data_type):
        """Load user data from Google Drive"""
        if not self.service:
            return self._local_load_data(username, data_type)
        
        # Implementation for Google Drive load
        pass
    
    def _local_create_folder(self, username):
        """Create local folder structure (fallback)"""
        base_path = Path(f"data/users/{username}")
        subfolders = ['assessments', 'results', 'coaching_sessions', 'uploads']
        
        for subfolder in subfolders:
            (base_path / subfolder).mkdir(parents=True, exist_ok=True)
        
        return str(base_path)
    
    def _local_save_data(self, username, data_type, data):
        """Save data locally (fallback)"""
        file_path = Path(f"data/users/{username}/{data_type}.json")
        file_path.parent.mkdir(parents=True, exist_ok=True)
        
        with open(file_path, 'w') as f:
            json.dump(data, f, indent=2)
        
        return str(file_path)
    
    def _local_load_data(self, username, data_type):
        """Load data locally (fallback)"""
        file_path = Path(f"data/users/{username}/{data_type}.json")
        
        if file_path.exists():
            with open(file_path, 'r') as f:
                return json.load(f)
        
        return None
