requirements.txt
streamlit==1.29.0
streamlit-authenticator==0.2.3
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0
openai==1.6.1
anthropic==0.8.1
google-generativeai==0.3.2
plotly==5.18.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
bcrypt==4.1.2
PyYAML==6.0.1
Pillow==10.1.0
